feat: implementar flujo de aprobación de préstamos y corregir registro de pagos
Backend:
- Corregir error 403 en PaymentController (cambio de permiso a LOAN:PAYMENT).
- Crear LoanApprovalService y LoanApprovalController para aprobar/rechazar préstamos.
- Modificar LoanService para crear préstamos en estado PENDING.
- Agregar campos paidAmount, paidDate, isVoided, voidedAt a entidades Payment y LoanSchedule.
- Agregar migraciones V13 y V15 para columnas faltantes en schemas de tenants.

Frontend:
- Actualizar tipos de Loan, Payment y LoanSchedule en types/index.ts.
- Agregar modales y botones de Aprobar/Rechazar en Loans.tsx y LoanDetail.tsx.
- Corregir etiquetas de métodos de pago en Payments.tsx.
- Agregar ítem "Por Aprobar" en Sidebar.tsx con validación de permisos.

+++ src/main/java/com/kredio/backend/service/DashboardService.java
@@ -49,7 +49,7 @@ public class DashboardService {
 
         // 3. Próximos vencimientos (7 días)
         BigDecimal upcomingAmount = scheduleRepository.sumExpectedPaymentsBetween(tenantId, today, nextWeek);
-        int upcomingCount = scheduleRepository.countPendingSchedulesBetween(tenantId, today, nextWeek);
+        int upcomingCount = (int) scheduleRepository.countPendingSchedulesBetween(tenantId, today, nextWeek);
 
         // 4. Lista detallada de próximos vencimientos
         List<DashboardStatsDTO.UpcomingPaymentDTO> upcomingPayments = buildUpcomingPayments(tenantId, today, nextWeek);
